Database somewhat implemented but the actual database isn't appearaing or storing data

+++ frontend/app/src/main/java/vinh/nguyen/app/database/Workout.kt
@@ -29,5 +29,5 @@ data class Workout(
     val id: Int = 0,
     val exercise: String,
     val reps: Int,
-    val length: Int
+    val length: Long
 )
